Change replacement method
